slack-integration/requirements.txt
slack-sdk>=3.23.0
google-auth>=2.23.0
requests>=2.31.0
